Generate expected proportional catch numbers for each gear.

The function returns a tibble of the expected relative catch number for each gear for the provided parameters. The gear values add up to 1, so for single gear fisheries, this is will return 1; therefore this is only useful in multi-gear fisheries.

Usage

blicc_expected_catches(blicc_rp)

Arguments

blicc_rp

A list of posterior draws, reference points with associated direction, the data object and expected lengths from blicc_ref_pts function.

Value

A tibble of expected catches and residuals for each gear and MCMC iteration

Details

As well as the expected catch, the fit residuals are returned as a diagnostic. For the mpd fit, the function returns a tibble with the number of rows equal to the number of gears. For the MCMC fit, it returns the number of MCMC iterations multiplied by the number of gears.

Examples

blicc_expected_catches(trgl_rp) |>
  dplyr::group_by(gear) |>
  dplyr::summarise(resid=mean(resid))
#> # A tibble: 2 × 2
#>   gear       resid
#>   <chr>      <dbl>
#> 1 Gillnet -0.00596
#> 2 Trawl    0.00596