Get expected BLICC selectivity, mortality, survival and length frequency

For a set of parameter values, returns the expected selectivity, mortality, survival, and catch in each length bin based on the BLICC model as a tibble.

Usage

blicc_get_expected(Linf, Galpha, Mk, Fk, Sm, blicc_ld)

Arguments

Linf

Maximum mean length fish in the population can grow to

Galpha

Alpha parameter for the Gamma probability density function that governs growth variability.

Mk

Natural mortality (time in units of the growth rate K)

Fk

Vector of fishing mortality (time in units of the growth rate K) to be multiplied by selectivity to get fishing mortality in each length bin

Sm

Vector of parameters for all the selectivity functions including mixture weights at the end of the vector

blicc_ld

A standard data list suitable for the model (see function blicc_dat)

Value

A tibble of expected values for each length bin for each gear

Details

The model and provided parameter values are used to calculate the expected values for selectivity, survival, relative population numbers, total mortality and expected length frequency for each length bin. Input parameters are point estimates. A standard data list provides the model dimensions as produced by the function blicc_dat.

Examples

blicc_get_expected(Linf = 62, Galpha = 100, Mk = 1.5, Fk = 1.5,
                   Sm = c(24, 0.1), blicc_ld = gillnet_ld)
#> # A tibble: 25 × 5
#>    Qgroup  Lgroup      sel    N_L     efq
#>    <chr>   <fct>     <dbl>  <dbl>   <dbl>
#>  1 Gillnet 14     0.000304 0.0419   0.268
#>  2 Gillnet 16     0.00745  0.0410   6.42 
#>  3 Gillnet 18     0.0821   0.0399  68.9  
#>  4 Gillnet 20     0.407    0.0382 327.   
#>  5 Gillnet 22     0.905    0.0354 673.   
#>  6 Gillnet 24     0.905    0.0319 608.   
#>  7 Gillnet 26     0.407    0.0293 251.   
#>  8 Gillnet 28     0.0821   0.0278  48.0  
#>  9 Gillnet 30     0.00745  0.0268   4.19 
#> 10 Gillnet 32     0.000304 0.0258   0.165
#> # ℹ 15 more rows