Impact from the removal of each gear on the YPR and SPR in the specified period.

The change in YPR and SPR are calculated for each parameter set to produce a tibble of the YPR and SPR change from the removal of each gear in sequence. This is only useful if there is more than one gear.

Usage

blicc_impact(blicc_rp)

Arguments

blicc_rp

A list of posterior draws, reference points with associated direction, the data object and expected lengths from blicc_ref_pts function.

Value

A tibble with the changes in YPR and SPR from the removal of each gear for the scenario in blicc result list.

Examples

blicc_impact(blicc_ref_pts(blicc_mpd(trgl_ld), trgl_ld))
#> Chain 1: Initial log joint probability = -1226.94
#> Chain 1:     Iter      log prob        ||dx||      ||grad||       alpha      alpha0  # evals  Notes 
#> Chain 1: Exception: neg_binomial_2_lpmf: Location parameter[1] is -nan, but must be positive finite! (in 'BLICC', line 291, column 8 to column 61)
#> 
#> Chain 1:      356      -177.932   2.97425e-06    0.00705275      0.2847           1      415   
#> Chain 1: Optimization terminated normally: 
#> Chain 1:   Convergence detected: relative gradient magnitude is below tolerance
#> # A tibble: 3 × 4
#>   .draw Variable    `Remove Trawl` `Remove Gillnet`
#>   <dbl> <chr>                <dbl>            <dbl>
#> 1     0 Trawl YPR           -0.504            0.219
#> 2     0 Gillnet YPR          0.108           -0.349
#> 3     0 SPR                  0.338            0.167