Generate posterior draws and standard reference points

The function returns a list consisting of the search direction vector, a tibble of draws from the posterior MCMC stanfit (or the mpd estimate) including SPR 20% and 40% reference points, expected number in each length bin and the data object used to generate these. The tibble of draws is suitable for use in the posterior and bayesplot packages and can be used in various functions to show results. Note that the function can take a little time to run and may be moderately large, depending on the number of draws. Some reference points may not exist (indicated by NA).

Usage

blicc_ref_pts(
  slimf,
  blicc_ld,
  time_period = NULL,
  vdir = NULL,
  a = NA,
  b = NA,
  L50 = NA,
  L95 = NA
)

Arguments

slimf

An object from the blicc_fit or blicc_mpd functions.

blicc_ld

A standard data list suitable for the model (see function blicc_dat)

time_period

Specifies the time period used for reference point scenario. If not specified, the default is the final time period.

vdir

A search direction vector with maximum value 1 and minimum 0 applied to changes across gears. Optional.

a

The length-weight parameter: a*L^b. Not used in model fitting, but used to calculate wt_L if that is not provided. for plots etc. Optional.

b

The length-weight exponent (a*L^b, usually close to 3.0). Not used in model fitting, but used to calculate wt_L if that is not provided. Optional.

L50

The length at 50% maturity, often referred to as "first" maturity, and primarily applies to females. Must be less than Linf, usually around 0.66 Linf, which is assumed if it is not provided. Optional.

L95

The length at 95% maturity. Must be greater than L50 and less than Linf. A small increment is added to L50 if it is not provided. Optional.

Value

A list of 1) dr_df: posterior draws or mpd point estimates of important parameters 2) lx_df: observed-expected length frequencies, 3) ld: blicc_ld data list used. 4) scenario used: includes a direction vector and subsets of the data and parameters sufficient to calculate rp_df 5) rp_df: per-recruit reference points

Details

The reference points calculated are F40 and F20 for 40% and 20% SPR respectively, and changes in the selectivity location parameter to achieve either SPR 40% (S40) or maximum yield based on yield-per-recruit (SMY). As well as reference points, the function recalculates the SPR if the length-weight or maturity parameters are provided. In addition, the exploitable biomass as a proportion of the unexploited biomass (B_B0) for each time period and the YPR for each data frequency are calculated for each parameter draw.

Reference points are calculated for a particular scenario, represented by a time period. By default, this time period is the last in the data set, although the period can be specified. This is necessary because different time periods can have different gears operating, which affects fishing mortality at length and therefore the fishing mortality reference points. This has no effect if there is only one time period.

Reference points are found where possible for the parameter draw from the MCMC. The reference points are found along a line representing the relative impact across gears. The default is to apply the same adjustments proportionately to every gear that has a fishing mortality, but alternative scenarios can be defined. For example, for two gears the default is vdir=c(1, 1). If there are two gears, and only the first will be adjusted, then the direction would be vdir=c(1, 0). It is possible to provide a vector such as vdir=c(1, 0.5) to represent only a 50% change in the second gear compared to the first, for example. This will only be applied to the fishing mortality. The selectivity changes will continue to use the vector as a dummy variable, but the quantity will be ignored and all vdir with non-zero values for gears will be treated as having equal displacement; so for example a selectivity evaluation for vdir = c(1.0, 0.5) is equivalent to vdir = c(1, 1).

It is likely that proper simulation projections would be required for more complex cases where gears are not being simply excluded from management action. Unless there is a specific need, the best option is to leave it as the default vdir=NULL so the reference points are indicative and consistent.

Where reference points do not exist an NA is returned. It is quite possible reference points do not exist in many cases due to the effects of selectivity. Strictly speaking, reference points may not exist within specific constraints on F and selectivity. Within more general bounds on F and selectivity, all reference "points" exist as lines or points. In these functions, reference points in relation to a control (Fk or Sm) are calculated based on the other being fixed at the current estimated value. To see how the fishery might respond when they are adjusted together, the yield or SPR surface must be plotted.

Examples

if (FALSE) { # \dontrun{
res_rp <- blicc_ref_pts(trgl_slim, trgl_ld)
summary(res_rp$rp_df)
} # }