Return a tibble containing a summary of the results from a fishblicc fit

This function provides a summary of results from the fitted model in a form suitable for displaying in a table. Note that other packages, such as rstan, bayesplot and posterior can extract information from the stanfit object and the blicc_ref_pts rp_df 'draws' object in more detailed form.

Usage

blicc_results(blicc_res)

Arguments

blicc_res

Results from blicc_fit (stanfit object), blicc_mpd or blicc_ref_pts

Value

A tibble summarising results of the model fit object.

Examples

blicc_results(trgl_slim)
#> # A tibble: 13 × 5
#>    Parameter       Mean        SD `N (eff)`  Rhat
#>    <chr>          <dbl>     <dbl>     <dbl> <dbl>
#>  1 Linf        66.3      1.74          745. 1.01 
#>  2 Galpha     109.      25.2           918. 1.00 
#>  3 Mk           1.54     0.148         746. 1.00 
#>  4 Fk[1]        0.457    0.106         725. 1.00 
#>  5 Fk[2]        0.766    0.175         670. 1.01 
#>  6 Sm[1]       16.1      0.415         911. 0.999
#>  7 Sm[2]        0.572    0.0578        817. 1.00 
#>  8 Sm[3]       34.2      0.393         714. 1.00 
#>  9 Sm[4]        0.00803  0.000442      792. 1.00 
#> 10 NB_phi     157.      71.8           869. 0.997
#> 11 Gbeta        1.64     0.364         913. 1.00 
#> 12 SPR[1]       0.314    0.0852        648. 1.00 
#> 13 lp__      -180.       2.11          659. 1.00