Sets data object from blicc_dat to have new selectivity functions

Checks selectivity functions are valid, then updates the data object with the new functions so that the parameters and parameter references for the selectivity functions are valid. New references are replaced in the data object, which is then returned.

Usage

blicc_selfun(blicc_ld, sel_fun, sel_indx = NULL, model_name = NULL)

Arguments

blicc_ld

A standard data list suitable for the model (see function blicc_dat)

sel_fun

A vector of selectivity function names (character) or index (1-4), with 1="logistic", 2="normal", 3="ss_normal", 4="ds_normal". Length must be the same as fq if gear_sel is not provided. Required.

sel_indx

An integer vector index of the selectivity functions to be changed. If not provided, all functions will be replaced.

model_name

A name for the model (species or fishery). Optional.

Value

The data object blicc_ld with the new selectivities

Details

Particular selectivity functions can be changed using the sel_indx parameter, which indexes the relevant functions. In this case, only those functions will need to have the prior parameters updated.

Examples

new_ld <- blicc_selfun(gillnet_ld, sel_fun="logistic", model_name = "Logistic Selectivity")