Set Fk log-normal prior hyper-parameters in the data object

The Fk lognormal prior is updated with new mu for each gear and a single sigma parameter. If no value (NA) is provided for lFk, the defaults are assigned. These are the current natural mortality for fishing mortality mu and 2.0 for the sigma parameter. The sigma hyper-parameter should be kept reasonably large to allow the model to estimate this parameter. In general, the default values should be sufficient as the prior should be non-informative.

Usage

blip_Fk(blicc_ld, lFk = NULL, lFks = NULL, model_name = NULL)

Arguments

blicc_ld

A standard data list suitable for the model (see function blicc_dat)

lFk

A vector containing the log-normal mean Fk for each gear having a non-negligible catch

lFks

A double containing the same log-normal sigma Fk for all gears

model_name

A string for a replacement model name in the data object. Optional.

Value

The data object blicc_ld with the prior for Fk changed.

Details

For the multigear model, if a gear has negligible catches so the relative catch vector indicates zero catch for that gear, no fishing mortality is estimated for it, and the gear's Fk prior is skipped as it is fixed at zero.

Examples

new_ld <- blip_Fk(gillnet_ld, lFk=log(1.9), lFks=1.5)