Set Galpha log-normal prior hyper-parameters in the data object

The Galpha lognormal prior is updated with new values for the mean (mu) and standard deviation (sigma). The CV is likely to be between 5% (Galpha mu = log(1/0.05^2)) and 30% (Galpha mu = log(1/0.3^2)) Values outside this range are not recommended. A 30% CV makes length very uninformative on age.

Usage

blip_Galpha(blicc_ld, lGalpha)

Arguments

blicc_ld

A standard data list suitable for the model (see function blicc_dat)

lGalpha

A numeric vector of double containing the mean and sd for the prior log-normal.

Value

The data object blicc_ld but with the prior for Galpha changed.

Details

Galpha is the alpha parameter in the gamma distribution for the Linf probability density. The default in blicc_dat is CV=10%: lGa1pha = c(log(1 / 0.1 ^ 2), 0.25).

Examples

new_ld <- blip_Galpha(gillnet_ld, lGalpha=c(log(1/0.05^2), 0.1))