Set life-history parameters and vectors in the data object

Sets the maturity-at-length and weight-at-length vectors in the data object (from blicc_dat) to new values. The values are either provided as vectors or model parameters to calculate the vectors. Calculations are done for the length bin mid-points. The models for the calculation are the standard length-weight (W=aL^b) and logistic model for the maturity-at-length: (Lm = 1/(1+exp(Ls*(L-L50)))).

Usage

blip_LH(
  blicc_ld,
  a = NULL,
  b = NULL,
  L50 = NULL,
  L95 = NULL,
  ma_L = NULL,
  wt_L = NULL,
  model_name = NULL,
  set_defaults = FALSE
)

Arguments

blicc_ld

A standard data list suitable for the model (see function blicc_dat)

a

The length-weight parameter: a*L^b. Not used in model fitting, but used to calculate wt_L if that is not provided. for plots etc. Optional.

b

The length-weight exponent (a*L^b, usually close to 3.0). Not used in model fitting, but used to calculate wt_L if that is not provided. Optional.

L50

The length at 50% maturity, often referred to as "first" maturity, and primarily applies to females. Must be less than Linf, usually around 0.66 Linf, which is assumed if it is not provided. Optional.

L95

The length at 95% maturity. Must be greater than L50 and less than Linf. A small increment is added to L50 if it is not provided. Optional.

ma_L

Mature biomass for each length bin. Optional.

wt_L

Weight (biomass) for each length bin. Optional.

model_name

A string for a replacement model name in the data object. Optional.

set_defaults

Logical indicating whether to set defaults if parameters are NA. Leaves them alone if FALSE.

Value

The data object blicc_ld with the new life-history parameters

Examples

new_ld <- blip_LH(gillnet_ld, a=1.2e-4, b=3, model_name="Alternative LW")
#> Warning: L50 or L95 not specified: mature biomass -at-length not changed.