Set the NB_phi prior hyper-parameters in the data object

The NB_phi log-normal prior is updated with a mean (mu) and standard deviation (sigma). NB_phi controls the negative binomial over-dispersion compared to the Poisson distribution. It is sometimes useful to reduce the NB_phi lognormal standard deviation hyper-parameter below the 0.5 default to avoid variation in the data being interpreted as observation error when it can be explained better by the model.

Usage

blip_NBphi(blicc_ld, lNBphi)

Arguments

blicc_ld

A standard data list suitable for the model (see function blicc_dat)

lNBphi

A numeric vector of double containing the mu and sigma for the prior log-normal.

Value

The data object blicc_ld but with the prior for NBphi changed.

Details

The variance for the negative binomial is Var = mu + mu^2 / NB_phi compared to Var = mu for the Poisson distribution.

Examples

new_ld <- blip_NBphi(gillnet_ld, lNBphi = c(log(100), 0.05))