Set a selectivity function's prior hyper-parameters to the given values

Each parameter set is for a single selectivity function indexed by sel_indx. The location parameter (loc) is either the 50% selectivity for the logistic or the mode of the particular normal selectivity function, and is required.

Usage

blip_sel(blicc_ld, sel_indx, loc, lslope = NULL)

Arguments

blicc_ld

A standard data list suitable for the model (see function blicc_dat)

sel_indx

Single integer indexing a selectivity function

loc

Single location parameter for the selectivity function

lslope

Vector of prior selectivity log slope parameters manually set for the specified selectivity function. Optional.

Value

The data object blicc_ld but with the new selectivity priors set

Details

The lslope parameter is a single or double log value of the slope for the relevant function. This is either the log of the logistic slope parameter, or the log of the reciprocal of the variance for the normal functions. For the double-sided normal, two parameters must be provided and for all other functions only one parameter. If lslope is not provided, a default slope of -4.5 is applied if the slope parameters have not already been set. If values are already present these are conserved. Mixture weights, if they are used, can be set in blip_mix.

Examples

new_ld <- blip_sel(gillnet_ld, sel_indx = 1, loc = 30, lslope = -2.8)