Example small sample of length frequency data from gillnet fishery

A data list created using simulated length frequency data from the function blicc_dat:

Usage

gillnet_ld

Format

A list of 26 data vectors used in the BLICC model:

model_name

Model and/or fit and/or species name

NQ

Number of length frequency data sets

NG

Number of gears

NS

Number of selectivity functions

NT

Number of time periods

NF

Number of fishing mortality estimates

NB

Number of bins in the length frequency

NP

Number of selectivity function parameters

NM

Number of mixture selectvity models

fqname

Length frequency data set names

gname

Fishing gear names

tpname

Time period names

LLB

Vector of length bin lower boundaries

LMP

Length bin mid-points, used for selectivity and plotting

fq

List of length frequency vectors, one for each gear

Gi

Integer vector linking length frequency data sets to gears

Ti

Integer vector linking length frequency data sets to time periods

prop_catch

Vector of proportional catch by gear of length NF

Fkq

Integer vector of length frequency data sets index linking to fishing mortality

fSel

Index of selectivity function to use for each selectvity model

GSbase

Index of based selectivity model for each gear

GSmix1

Index of start and end mixtures in GSmix2: 0 implies no mixture

GSmix2

Index of all selectivity mixture models referenced by gear in GSmix1

poLinfm

Linf: Mean prior normal parameter

poLinfs

Linf: sd prior normal parameter

polGam

Galpha: Mean prior normal log parameter

polGas

Galpha: sd prior normal log parameter

a

Length-weight parameter in a L^b. Optional

b

Length-weight parameter in a L^b.

L50

Logistic parameter for 50% maturity

Ls

Logistic steepness parameter for the maturity ogive

wt_L

Vector of average weight in each length bin

ma_L

Vector of mature biomass in each length bin

polMkm

Mk: Mean prior normal log parameter

polMks

Mk: sd prior normal log parameter

polFkm

Fk: Mean prior normal log parameter

polFks

Fk: sd prior normal log parameter

sp_i

Integer index of the parameter start in polSm for each selectivity function

sp_e

Integer index of the parameter end in polSm for each selectivity function

polSm

Sm: Mean priors normal log selectivity parameter including mixture weights at the end

polSs

Sm: sd priors normal log selectivity parameter including mixture weights at the end

polNB_phim

NB_phi: Mean prior log normal parameter

polNB_phis

NB_phi: sd prior log normal parameter

polCs

Catch: sd prior log normal parameter on relative catch

NK

Number of nodes in the Gauss-Laguerre quadrature

gl_nodes

Gauss-Laguerre quadrature nodes

gl_weights

Gauss-Laguerre quadrature weights

Source

Data are simulated.

Details

gillnet_ld <- fishblicc::blicc_dat(model_name = "Simulated Gillnet Data", LLB = seq(14, 62, by=2), fq=list(Gillnet = c(0, 2, 7, 18, 32, 56, 115, 155, 227, 240, 246, 249, 210, 158, 106, 73, 43, 28, 7, 6, 4, 3, 0, 1, 0)), sel_fun = "normal", gear_names = "Gillnet", Linf=c(64, 3), Mk=1.55, a=1.0e-4, b=2.95, L50=34, L95=43)