Get a posterior predicted length frequency

For a set of MCMC parameter draws, returns a simulated catch in each length bin based on the BLICC model. This matrix of simulated data that might be expected if the model is correct, can be used in the posterior and bayesplot packages for evaluation.

Usage

posterior_predict(blicc_rp, gear = NULL, time_period = NULL, draws = 0)

Arguments

blicc_rp

List of fishblicc result tables from blicc_ref_pts

gear

Identifies the single gear (selectivity) providing predictions

time_period

Identifies the single time_period used to simulate a length frequency

draws

The number of random draws up to the number of draws from the MCMC (the default).

Value

A matrix with rows equal to gears*draws and columns to length

Examples

yrep <- posterior_predict(blicc_rp = trgl_rp, gear=1, time_period = 1, draws=100)