A list of results including reference points from the blicc_ref_pts function

An example data table of parameter draws and reference points from the BLICC model trgl_rp <- blicc_ref_pts(trgl_slim, trgl_ld)

Usage

trgl_rp

Format

A list of data frames of results and input information for a BLICC model fit:

dr_df

A draws tibble of parameter estimates

lx_df

A tibble of expected numbers of fish in each length bin for each gear

ld

The data list from blicc_dat used to fit the model and calculate indicators and results

scenario

A list of time period, gears, a vector describing the "direction" across gears and data list used to calculate reference points

rp_df

A draws tibble of parameter estimates and reference points for the scenario

rp_df is a tibble of parameter estimates from a BLICC model fit:

Linf

Mean maximum length from the von Bertalanffy growth model

Galpha

Gamma distribution parameter governing growth variability

Mk

Natural mortality (per unit K time)

Fk

A list vector of fishing mortality (per unit K time)

Sm

A list vector of selectivity model parameters

NB_phi

Negative binomial parameter: Excess variance compared to the Poisson

Gbeta

Gamma distribution "rate" parameter: (=Galpha/Linf)

.chain

MCMC chain identifier

.iteration

MCMC iteration identifier

.draw

MCMC draw identifier

F20

List vector of fishing mortality estimated to achieve 20% SPR

F40

List vector of fishing mortality estimated to achieve 40% SPR

S40

List vector of selectivity location estimated to achieve 40% SPR

SMY

List vector of selectivity location estimated to achieve maximum yield per recruit

Source

Data are simulated.