Example stanfit object from the BLICC model (Stan MCMC)

A stanfit object from fitting the BLICC model using blicc_fit to the "eg_ld" data set. The draws can be extracted and reference points calculated using function blicc_ref_pts. The MCMC simulations are also accessible through a variety of rstan and posterior functions. trgl_slim <- blicc_fit(trgl_ld, nwarmup=500, thin=2)

Usage

trgl_slim

Format

A stanfit object: variables can be obtained using rstan::extract and other methods. see packages rstan and posterior for details.

Source

Data are simulated.